requirements.txt
# Core camera and video processing
opencv-python==4.8.1.78
numpy==1.24.3

# Image processing
Pillow==10.1.0

# Motion detection utilities
imutils==0.5.4

# Configuration
pyyaml==6.0.1

# Development tools (optional)
ipython==8.17.2
black==23.11.0
pylint==3.0.2

